requirements.txt
streamlit
openai
pypdf
python-docx
python-dotenv
pytesseract
pdf2image
pillow